[common/aggregate] refine the tests
