+++ common/aggregate_functions/src/aggregate_batch_test.rs
@@ -11,19 +11,8 @@ use pretty_assertions::assert_eq;
 
 use crate::*;
 
-// Copyright 2020-2021 The Datafuse Authors.
-//
-// SPDX-License-Identifier: Apache-2.0.
-
 #[test]
-fn test_array_common_aggregate() {
-    use std::sync::Arc;
-
-    use pretty_assertions::assert_eq;
-
-    use crate::*;
-
-    #[allow(dead_code)]
+fn test_aggregate_batch() {
     struct ArrayTest {
         name: &'static str,
         args: Vec<ArrayRef>,
@@ -132,170 +121,9 @@ fn test_array_common_aggregate() {
         for (i, args) in t.args.iter().enumerate() {
             let result = (*(t.func))(args.clone().into());
             match result {
-                Ok(v) => assert_eq!(v, t.expect[i]),
+                Ok(v) => assert_eq!(v, t.expect[i], "{}", t.name),
                 Err(e) => assert_eq!(t.error[i], e.to_string()),
             }
         }
     }
-}
-
-#[test]
-fn test_aggregate_function() -> Result<()> {
-    #[allow(dead_code)]
-    struct Test {
-        name: &'static str,
-        eval_nums: usize,
-        args: Vec<DataField>,
-        display: &'static str,
-        nullable: bool,
-        columns: Vec<DataColumnarValue>,
-        expect: DataValue,
-        error: &'static str,
-        func_name: &'static str,
-    }
-
-    let columns: Vec<DataColumnarValue> = vec![
-        Arc::new(Int64Array::from(vec![4, 3, 2, 1])).into(),
-        Arc::new(Int64Array::from(vec![1, 2, 3, 4])).into(),
-    ];
-
-    let args = vec![
-        DataField::new("a", DataType::Int64, false),
-        DataField::new("b", DataType::Int64, false),
-    ];
-
-    let tests = vec![
-        Test {
-            name: "count-passed",
-            eval_nums: 1,
-            args: vec![args[0].clone()],
-            display: "count",
-            nullable: false,
-            func_name: "count",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::UInt64(Some(4)),
-            error: "",
-        },
-        Test {
-            name: "max-passed",
-            eval_nums: 2,
-            args: vec![args[0].clone()],
-            display: "max",
-            nullable: false,
-            func_name: "max",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::Int64(Some(4)),
-            error: "",
-        },
-        Test {
-            name: "min-passed",
-            eval_nums: 2,
-            args: vec![args[0].clone()],
-            display: "min",
-            nullable: false,
-            func_name: "min",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::Int64(Some(1)),
-            error: "",
-        },
-        Test {
-            name: "avg-passed",
-            eval_nums: 1,
-            args: vec![args[0].clone()],
-            display: "avg",
-            nullable: false,
-            func_name: "avg",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::Float64(Some(2.5)),
-            error: "",
-        },
-        Test {
-            name: "sum-passed",
-            eval_nums: 1,
-            args: vec![args[0].clone()],
-            display: "sum",
-            nullable: false,
-            func_name: "sum",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::Int64(Some(10)),
-            error: "",
-        },
-        Test {
-            name: "argMax-passed",
-            eval_nums: 1,
-            args: args.clone(),
-            display: "argmax",
-            nullable: false,
-            func_name: "argmax",
-            columns: columns.clone(),
-            expect: DataValue::Int64(Some(1)),
-            error: "",
-        },
-        Test {
-            name: "argMin-passed",
-            eval_nums: 1,
-            args: args.clone(),
-            display: "argmin",
-            nullable: false,
-            func_name: "argmin",
-            columns: columns.clone(),
-            expect: DataValue::Int64(Some(4)),
-            error: "",
-        },
-        Test {
-            name: "argMin-notpassed",
-            eval_nums: 1,
-            args: vec![args[0].clone()],
-            display: "argmin",
-            nullable: false,
-            func_name: "argmin",
-            columns: columns.clone(),
-            expect: DataValue::Int64(Some(4)),
-            error: "Code: 28, displayText = argmin expect to have two arguments, but got 1.",
-        },
-        Test {
-            name: "uniq-passed",
-            eval_nums: 1,
-            args: vec![args[0].clone()],
-            display: "uniq",
-            nullable: false,
-            func_name: "uniq",
-            columns: vec![columns[0].clone()],
-            expect: DataValue::UInt64(Some(4)),
-            error: "",
-        },
-    ];
-
-    for t in tests {
-        let rows = t.columns[0].len();
-
-        let func = || -> Result<()> {
-            let mut func1 = AggregateFunctionFactory::get(t.func_name, t.args.clone())?;
-
-            for _ in 0..t.eval_nums {
-                func1.accumulate(&t.columns, rows)?;
-            }
-            let state1 = func1.accumulate_result()?;
-
-            let mut func2 = AggregateFunctionFactory::get(t.func_name, t.args.clone())?;
-            for _ in 1..t.eval_nums {
-                func2.accumulate(&t.columns, rows)?;
-            }
-            let state2 = func2.accumulate_result()?;
-
-            let mut final_func = AggregateFunctionFactory::get(t.func_name, t.args.clone())?;
-            final_func.merge(&*state1)?;
-            final_func.merge(&*state2)?;
-
-            let result = final_func.merge_result()?;
-            assert_eq!(&t.expect, &result);
-            assert_eq!(t.display, format!("{:}", final_func));
-            Ok(())
-        };
-
-        if let Err(e) = func() {
-            assert_eq!(t.error, e.to_string());
-        }
-    }
-    Ok(())
 }
